more work on vss
