+++ node/src/vsapi/vs.thrift.sum
@@ -1,0 +1,1 @@
+1bb0467e84058f40ba4e3a44aa2f2ca6  vs.thrift
