+++ node/src/vssapi/vss.thrift.sum
@@ -1,0 +1,1 @@
+750007ec5ad11af69cd6074ab5a4a883  vss.thrift
