+++ visaservice/core/pkg/vservice/vsinst.go
@@ -2,29 +2,22 @@ package vservice
 
 import (
 	"context"
-	"crypto/md5"
 	"crypto/tls"
 	"errors"
-	"fmt"
-	"math/rand" // not used for crypto
+	"fmt" // not used for crypto
 	"net/netip"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/apache/thrift/lib/go/thrift"
-	"google.golang.org/protobuf/proto"
 
 	"zpr.org/vs/pkg/agent"
 	snip "zpr.org/vs/pkg/ip"
-	"zpr.org/vs/pkg/libvisa"
 	"zpr.org/vs/pkg/logr"
 	"zpr.org/vs/pkg/policy"
-	"zpr.org/vs/pkg/snauth"
 	"zpr.org/vs/pkg/vsapi"
 	"zpr.org/vs/pkg/vservice/auth"
 	"zpr.org/vsx/snio/vsio"
-	"zpr.org/vsx/snio/zds"
 )
 
 var (
@@ -34,29 +27,32 @@ var (
 	ErrAuthExpired    = errors.New("auth expired")
 )
 
-// The visa-service "peers" are always nodes.
-type PeerRecord struct {
-	ZPRAddr              netip.Addr // can use to lookup in agentDB
-	RegistrationTime     time.Time
-	LastPollTime         time.Time
-	VisaRequestsCount    uint64
-	ConnectRequestsCount uint64
-}
-
 type HelloRecord struct {
 	CTime  time.Time
 	Chksum uint32
 }
 
+// All agents in the system have a HostRecord.
+// Nodes will have the Peer struct set.
 type HostRecord struct {
 	CTime        time.Time // connect/create time
 	LastAuthTime time.Time
 	Agent        *agent.Agent
 	ZPRAddr      netip.Addr
 	TetherAddr   netip.Addr
-	APIKey       string
-	VSSAddr      string
-	VSSState     VSSStateT
+	Peer         *PeerRecord
+}
+
+// The visa-service "peers" are always nodes.
+type PeerRecord struct {
+	APIKey               string
+	RegistrationTime     time.Time
+	LastPollTime         time.Time
+	VisaRequestsCount    uint64
+	ConnectRequestsCount uint64
+	VSSAddr              string
+	VSSState             VSSStateT
+	PushBuffer           []*vsapi.PollResponse
 }
 
 type VSSStateT int
@@ -78,6 +74,12 @@ type VSMsg struct {
 	NodeAddr netip.Addr
 }
 
+type PushItem struct {
+	Broadcast bool
+	NodeAddr  netip.Addr
+	Item      *vsapi.PollResponse
+}
+
 // VSInst is an instance of distributed visa service
 //
 // This is a bit of a mess at the moment as we are in progress of porting this from
@@ -88,17 +90,15 @@ type VSInst struct {
 	hopCount             uint
 	authr                auth.AuthService
 	attrProx             *AttrProxy
-	visaPushC            chan *vsapi.PollResponse // For pushing visas without needing a request
+	visaPushC            chan *PushItem // For pushing visas without needing a request
 	nodeNumber           uint8
 	nodeState            ConstraintService
 	thriftServer         thrift.TServer
 	vsMsgC               chan *VSMsg
-	vssCli               *VSSCli
 	localAddr            netip.Addr
 	thriftWg             sync.WaitGroup
 	thriftCreds          *tls.Config
 	exitC                chan struct{}
-	mb                   *Mailbox
 	reauthBumpTime       time.Duration
 	accessToken          []byte // Access token for special node operations
 	allowInvalidPeerAddr bool   // Set to TRUE for testing only.
@@ -129,7 +129,7 @@ type VSInst struct {
 	sessions struct {
 		sync.RWMutex
 		hellos  map[int32]*HelloRecord
-		apiKeys map[string]*PeerRecord
+		apiKeys map[string]netip.Addr // ZPR Addr (can use to lookup in the agent DB)
 	}
 }
 
@@ -163,9 +163,8 @@ func NewVSInst(vcf *VSIConfig) (*VSInst, error) {
 	vs := &VSInst{
 		log:                  vcf.Log,
 		hopCount:             vcf.HopCount,
-		visaPushC:            make(chan *vsapi.PollResponse, 128), // Must be large enough to handle a mass revocation event
+		visaPushC:            make(chan *PushItem, 128), // Must be large enough to handle a mass revocation event
 		thriftCreds:          vcf.Creds,
-		mb:                   NewMailbox(vcf.Log),
 		reauthBumpTime:       DefaultReauthBumpTime,
 		exitC:                make(chan struct{}),
 		accessToken:          vcf.AccessToken,
@@ -178,7 +177,7 @@ func NewVSInst(vcf *VSIConfig) (*VSInst, error) {
 	}
 	vs.vtable.table = make(map[uint32]*vtableEnt)
 	vs.vtable.nextVisaID = minVisaID
-	vs.sessions.apiKeys = make(map[string]*PeerRecord)
+	vs.sessions.apiKeys = make(map[string]netip.Addr)
 	vs.sessions.hellos = make(map[int32]*HelloRecord)
 	vs.agentDB.agents = make(map[netip.Addr]*HostRecord)
 
@@ -245,11 +244,8 @@ VS_RUNLOOP:
 			vs.periodicHousekeeping(now)
 
 		case req, ok := <-vs.visaPushC: // Drain this push channel
-			// TODO: This is not a great plan. These boxes for the docks and forwarders could get large.
-			//       Also, polling is not a good fit for the revoke use-case. In that case we want to
-			//       kill the visa/credential immediately.
 			if ok {
-				vs.mb.AppendMessage(req)
+				vs.pushToNode(req)
 			}
 
 		case <-vs.exitC:
@@ -267,21 +263,6 @@ func (vs *VSInst) Stop() {
 	close(vs.exitC) // stop runloop
 }
 
-func (vs *VSInst) HasPollerNode(nodeAddr netip.Addr) bool {
-	return vs.mb.HasPoller(nodeAddr.String())
-}
-
-func (vs *VSInst) HasRegisteredNode(nodeAddr netip.Addr) bool {
-	vs.sessions.RLock()
-	defer vs.sessions.RUnlock()
-	for _, pr := range vs.sessions.apiKeys {
-		if nodeAddr == pr.ZPRAddr {
-			return true
-		}
-	}
-	return false
-}
-
 // Implement policy.Configurator interface.
 // TODO: Pretty sure this is irrelevant for the visa service.  Was (is?) used to alter some configuration values
 // on a node.  For now logging when this is used to see if we need it.
@@ -296,78 +277,12 @@ func (vs *VSInst) periodicHousekeeping(now time.Time) {
 	vs.removeExpiredVisas()
 	vs.expireOldConfiguration()
 	vs.checkNodesVSSState()
+	vs.checkPushBuffers()
 }
 
 // RunPeriodicHousekeepingNow is here for unit tests only. Do not call outside of unit tests.
 func (vs *VSInst) RunPeriodicHousekeepingNow() {
 	vs.periodicHousekeeping(time.Now())
-}
-
-func (vs *VSInst) handleNodeRegister(nodeAddr netip.Addr) {
-	vs.log.Info("node registered", "nodeAddr", nodeAddr)
-	// TODO: More here in the future.
-	//
-	// Initially we were calling into the VSS here, but the node has not started the VSS
-	// at registration time -- it starts it after successful registration.
-}
-
-// checkNodeVSSState checks the VSS state of all nodes and sends config and policy to nodes
-// which indicate they are uninitialized.
-func (vs *VSInst) checkNodesVSSState() {
-	var nodes []netip.Addr
-	vs.agentDB.Lock()
-	for nodeAddr, rec := range vs.agentDB.agents {
-		if rec.VSSState == VSSStateUninitialized {
-			nodes = append(nodes, nodeAddr)
-		}
-	}
-	vs.agentDB.Unlock()
-
-	for _, nodeAddr := range nodes {
-		vs.sendConfigAndPolicyToNode(nodeAddr)
-	}
-}
-
-func (vs *VSInst) sendConfigAndPolicyToNode(nodeAddr netip.Addr) {
-	var serviceAddr string
-	var vssPrevState VSSStateT
-
-	vs.agentDB.Lock()
-	if rec, ok := vs.agentDB.agents[nodeAddr]; ok {
-		serviceAddr = rec.VSSAddr
-		vssPrevState = rec.VSSState
-		if serviceAddr != "" {
-			if vssPrevState == VSSStateUninitialized {
-				rec.VSSState = VSSStateInitializing
-			}
-		}
-	}
-	vs.agentDB.Unlock()
-
-	if serviceAddr == "" {
-		vs.log.Warn("node registered but VSS service address not in agentDB", "nodeAddr", nodeAddr)
-		return
-	}
-
-	if vssPrevState != VSSStateUninitialized {
-		// nothing to do
-		return
-	}
-
-	var vssNextState VSSStateT
-	plcy, _, cid := vs.getPolicyMatcherConfig()
-	if err := vs.vssCli.SendNetworkPolicy(serviceAddr, plcy.VersionNumber(), cid); err != nil {
-		vs.log.WithError(err).Error("failed to send network policy message to node", "service_addr", serviceAddr)
-		vssNextState = VSSStateUninitialized
-	} else {
-		vssNextState = VSSStateInitialized
-	}
-
-	vs.agentDB.Lock()
-	if rec, ok := vs.agentDB.agents[nodeAddr]; ok {
-		rec.VSSState = vssNextState
-	}
-	vs.agentDB.Unlock()
 }
 
 // expireOldConfiguration expires the oldest configuration change which has exceeded
@@ -425,27 +340,6 @@ func (vs *VSInst) extendVisaServiceVisas() {
 	vs.rerequestVisas(expiringVisas, (2 * VSVisaRenewalTime), true, vs.plcy.p.VersionNumber())
 }
 
-// renewEssentialVisasForCurrentConfig renews the visa-service visas found in our
-// store if they are not already on the current net-config.
-func (vs *VSInst) renewEssentialVisasForCurrentConfig(configID, policyID uint64) {
-	var oldVisas []*vtableEnt
-
-	vs.vtable.mtx.RLock()
-	for _, ve := range vs.vtable.table {
-		if ve.isVSVisa {
-			if ve.v.Configuration != configID && ve.successor == 0 {
-				oldVisas = append(oldVisas, ve)
-			}
-		}
-	}
-	vs.vtable.mtx.RUnlock()
-	if sz := len(oldVisas); sz > 0 {
-		vs.log.Info("re-requesting essential visas due to config change", "count", sz)
-	}
-	// Create a new visa with same parameters as original, and push to nodes.
-	vs.rerequestVisas(oldVisas, (2 * VSVisaRenewalTime), true, policyID)
-}
-
 // rerequestVisas requests "successor" visas for the visas in the passed list.
 func (vs *VSInst) rerequestVisas(xvisas []*vtableEnt, minDuration time.Duration, push bool, expectedPolicyID uint64) {
 	for _, ve := range xvisas {
@@ -463,7 +357,10 @@ func (vs *VSInst) rerequestVisas(xvisas []*vtableEnt, minDuration time.Duration,
 			// vs.dumpVisaTableHoldingLock("rerequest")
 			vs.vtable.mtx.Unlock()
 			if push {
-				vs.mb.AppendVisaResponseMessage(resp)
+				// TODO: To push the visa we need to know which nodes need this.
+				//       I think we used to put this in the mailbox for all nodes,
+				//       for now I am broadcasting.
+				vs.PushVisaToAllNodes([]*vsapi.VisaHop{resp.Visa})
 			}
 		}
 	}
@@ -482,15 +379,6 @@ func (vs *VSInst) removeExpiredVisas() {
 	// vs.dumpVisaTableHoldingLock("removeExpired")
 }
 
-// Push the visa for polling nodes.
-//
-// TODO: Presumably the mailbox system makes sure the right visa goes to the right node?
-//
-// Note blocks if our push channel is full.
-func (vs *VSInst) PushVisa(pr *vsapi.PollResponse) {
-	vs.visaPushC <- pr
-}
-
 // expireAllVisas is called when policy is updated.  Revokes and removes all visas under the
 // given network configuration.
 func (vs *VSInst) expireAllVisas(config uint64) {
@@ -499,18 +387,20 @@ func (vs *VSInst) expireAllVisas(config uint64) {
 	count := 0
 	for vID, ve := range vs.vtable.table {
 		if ve.v.Configuration == config {
-			revoke := &vsapi.PollResponse{
-				Revocations: []*vsapi.VisaRevocation{
-					{
-						IssuerID:      int32(vID),
-						Configuration: int64(config),
+			item := &PushItem{
+				Broadcast: true,
+				Item: &vsapi.PollResponse{
+					Revocations: []*vsapi.VisaRevocation{
+						{
+							IssuerID:      int32(vID),
+							Configuration: int64(config),
+						},
 					},
 				},
 			}
-
 			// We are often called from runloop so blocking here would be bad.
 			select {
-			case vs.visaPushC <- revoke: // ok
+			case vs.visaPushC <- item: // ok
 			default:
 				vs.log.Warn("push channel full, failed to issue revoke, continuing")
 			}
@@ -519,482 +409,4 @@ func (vs *VSInst) expireAllVisas(config uint64) {
 		}
 	}
 	vs.log.Infof("%d visas revoked due to configuration change", count)
-}
-
-// RequestVisa perform a visa request operation.
-// Set `minDuration` to force a minimum TTL on the visa, or set to 0 to leave
-// calculated expiration value alone.
-//
-// TODO: We do not pay attention to the context. If the context expires the
-//
-//	caller (dock, for example) will ignore the response.
-func (vs *VSInst) doRequestVisa(ctx context.Context, tetherAddr netip.Addr, pktData *snip.Traffic, minDuration time.Duration, expectedPolicyID uint64) (*vsapi.VisaResponse, error) {
-	// Packet will either be an opening request of a client to a service, or a response
-	// from a service to a client.  The addresses in the packet will be contact addresses.
-	//
-	// We need to see if there is any policy on file that permits this communication and if so, issue a visa.
-	//
-	// The visa will be expressed in tether addresses. But the PEP args will include contact addresses.
-	//
-
-	vs.log.Debug("RequestVisa starts", "zprSRC", pktData.SrcAddr, "zprDEST", pktData.DstAddr, "dport", pktData.DstPort)
-
-	curpol, curmatcher, curConfigID := vs.getPolicyMatcherConfig()
-	if curpol == nil || curmatcher == nil || curpol.IsEmpty() {
-		vs.log.Info("visa denied: nil or empty policy", "source", pktData.SrcAddr)
-		return nil, ErrDeniedByPolicy
-	}
-	if curpol.VersionNumber() != expectedPolicyID {
-		vs.log.Info("visa denied: version mismatch", "found", curpol.VersionNumber(), "expected", expectedPolicyID)
-		return nil, ErrVSMisconfigure
-	}
-
-	srcAgent, dstAgent, err := vs.endpointsForTraffic(pktData)
-	if err != nil {
-		return nil, err
-	}
-
-	// Do not issue a visa if either of the agents has expired.
-	//
-	// Note that we allow the expiration to be ZERO to handle case where we are
-	// talking about the nodes internal tunnel.  Possibly it would be better to
-	// actually set an expire time out our internal tunnel.  In any case I could
-	// see wanting to actually re-auth the internal tunnel, in case data source
-	// values have changed, for example.
-	srcAgentExpire, dstAgentExpire := srcAgent.GetAuthExpires(), dstAgent.GetAuthExpires()
-	{
-		now := time.Now()
-		if (!srcAgentExpire.IsZero()) && now.After(srcAgentExpire) {
-			vs.log.Info("visa denied, source agent auth has expired")
-			return nil, ErrAuthExpired
-		}
-		if (!dstAgentExpire.IsZero()) && now.After(dstAgentExpire) {
-			vs.log.Info("visa denied, dest agent auth has expired")
-			return nil, ErrAuthExpired
-		}
-	}
-
-	dstTether := dstAgent.GetTetherAddr() // The source tether address is in passed in.
-	if !dstTether.IsValid() {
-		vs.log.Info("destination tether is nil, visa request denied")
-		return nil, ErrNoRouteToHost
-	}
-
-	if len(dstAgent.GetProvides())+len(srcAgent.GetProvides()) == 0 {
-		vs.log.Info("visa denied: no services offered on either endpoint")
-		return nil, ErrDeniedByPolicy
-	}
-
-	// See if the traffic matches a policy.  Note that the policy has many scopes.
-	// If it is ok, create a DOCK PEP and VISA that can be used to forward traffic like it.
-	//
-	// TODO: One problem with the current matching method is that it requires the full list of
-	//       agent attributes.  I shouldn't have to send those between visa services.  A visa
-	//       service can query the data sources directly for attrs.  But we do need a way to
-	//       identify agents between visa servics.  So if agentX connects at nodeY, when nodeZ
-	//       wants to request a visa, it needs to know it is talking about agentX.
-	//
-	//       Ah ha, maybe just share the agent IDENTITY credentials.  Those the the keys in the
-	//       new datasource API anyway.
-	//
-	now := time.Now()
-	{
-		updated, newAttrs, err := vs.checkAndUpdateAttrs(now, srcAgent)
-		if updated {
-			vs.log.Debug("found updates to source authed claims", "agent_addr", srcAgent.GetZPRIDIfSet(), "newAttrs", newAttrs)
-			srcAgent.SetAuthedClaims(newAttrs)
-		}
-		if err != nil {
-			if errors.Is(err, auth.ErrNotSupported) {
-				vs.log.Info("attribute query not supported for source agent", "agent", srcAgent.GetIdentity())
-			} else {
-				vs.log.WithError(err).Warn("attribute query failed for source agent", "agent", srcAgent.GetIdentity())
-			}
-		}
-	}
-	{
-		updated, newAttrs, err := vs.checkAndUpdateAttrs(now, dstAgent)
-		if updated {
-			vs.log.Debug("found updates to dest authed claims", "agent_addr", srcAgent.GetZPRIDIfSet(), "newAttrs", newAttrs)
-			dstAgent.SetAuthedClaims(newAttrs)
-		}
-		if err != nil {
-			if errors.Is(err, auth.ErrNotSupported) {
-				vs.log.Info("attribute query not supported for dest agent", "agent", srcAgent.GetIdentity())
-			} else {
-				vs.log.WithError(err).Warn("attribute query failed for dest agent", "agent", dstAgent.GetIdentity())
-			}
-		}
-	}
-
-	mtSrc, mtDst := policyAgentInfoFromAgent(srcAgent), policyAgentInfoFromAgent(dstAgent)
-	cpols, err := curmatcher.MatchTraffic(pktData, mtSrc, mtDst)
-	if err != nil {
-		vs.visaDenied(curConfigID, "no match", pktData, tetherAddr)
-		vs.log.WithError(err).Info("visa denied: match failed")
-		return nil, ErrDeniedByPolicy
-	}
-
-	// We set a temporary ID on it, giving it a final ID when we add it into our table.
-	builder := libvisa.NewVisaBuilder(curConfigID, tetherAddr, dstTether).WithIssuerID(1).
-		WithTrafficAndPolicy(pktData, cpols).
-		WithDatacapComputeFunc(vs.dataCapApply)
-
-	if cpols[0].FWD {
-		builder.WithClientAgentIdent(srcAgent.GetIdentity())
-	} else {
-		builder.WithClientAgentIdent(dstAgent.GetIdentity())
-	}
-
-	// In order to compute expiration I need two things from the visaConfig:
-	//  1. The Lifetime value (if any -- this is from a duration constraint)
-	//  2. The Cap "period" - a time.Duration
-	//
-	durationCons := libvisa.MaxDurationConstraintFromPolicies(cpols)
-
-	// For a given set of polcies there may be a single DataCap that applies.
-	// If so, grab the period from that datacap to use in our expiration calculations.
-	var dataCapPeriod time.Duration
-	if cap := libvisa.MaximalDataCapFromPolicies(cpols); cap != nil {
-		dataCapPeriod = cap.CapPeriod
-	}
-
-	// visaExpiration, expFlags := vs.computeVisaExpiration(curpol.GetMaxVisaLifetime(), visaConfig, srcAgentExpire, dstAgentExpire)
-	visaExpiration, expFlags := vs.computeVisaExpiration(curpol.GetMaxVisaLifetime(), durationCons, dataCapPeriod, srcAgentExpire, dstAgentExpire)
-	if (minDuration > 0) && time.Until(visaExpiration) < minDuration {
-		visaExpiration = time.Now().Add(minDuration)
-		expFlags |= libvisa.ExpFMinDur
-	}
-
-	// What do we do here?
-	if time.Now().After(visaExpiration) {
-		return nil, fmt.Errorf("unable to compute valid expiration (%v)", visaExpiration)
-	}
-
-	builder = builder.WithExpiration(visaExpiration)
-
-	sKey := make([]byte, 16)
-	snauth.NewNonce(sKey)
-	builder = builder.WithSessionKeyAndEncoding(sKey, libvisa.SKEv1)
-
-	visa, err := builder.Visa()
-	if err != nil {
-		return nil, fmt.Errorf("failed to create visa: %w", err)
-	}
-
-	// The visa service keeps track of all visas outstanding. Before returning this visa we insert it
-	// into our visa table, which generates an ID as a side effect.
-	isVSVisa := (vs.localAddr == pktData.DstAddr) && (VisaServicePort == pktData.DstPort)
-
-	vent, err := vs.insertVisaWithNewID(visa, isVSVisa, pktData)
-	if err != nil {
-		return nil, fmt.Errorf("failed to insert visa into table: %w", err)
-	}
-
-	// TODO: Sign visa
-
-	resp := new(vsapi.VisaResponse)
-	resp.Status = vsapi.StatusCode_SUCCESS
-
-	pbuf, err := proto.Marshal(vent.v)
-	if err != nil {
-		vs.log.WithError(err).Error("failed to marshal visa for mailbox")
-		return nil, fmt.Errorf("internal error")
-	}
-	resp.Visa = &vsapi.VisaHop{
-		VisaPb:   pbuf,
-		HopCount: int32(vs.hopCount),
-		IssuerID: int32(vent.v.IssuerId),
-	}
-
-	vs.visaCreated(vent.v, visaExpiration, pktData, expFlags.String(), tetherAddr)
-	if time.Until(visaExpiration) < (30 * time.Second) {
-		vs.log.Warn("visa with very short TTL", "visaID", vent.v.IssuerId, "TTL", time.Until(visaExpiration).String())
-	}
-	return resp, nil
-}
-
-// Urg, so many types!!
-func policyAgentInfoFromAgent(agnt *agent.Agent) *policy.AgentInfo {
-	aa := &policy.AgentInfo{
-		AgentAttrs:    make(map[string]*agent.ClaimV),
-		AgentProvides: agnt.GetProvides(),
-	}
-	for key, claim := range agnt.GetAuthedClaims() {
-		aa.AgentAttrs[key] = claim
-	}
-	return aa
-}
-
-// dataCapApply will track the application of a (possibly grouped) data cap.
-// Returns the key under which the DataCap is stored, and the amount of data (in bytes) remaining.
-//
-// This matches the interface require by the libvisa builder.
-//
-// TODO: Not sure how to safely clean out cap table.
-func (vs *VSInst) dataCapApply(fwd bool, cap *libvisa.DataCap, clientAgentIdent string) (capKey string, remain uint64, err error) {
-	capID := cap.SvcID
-	if cap.CapGroup != "" {
-		capID = cap.CapGroup
-	}
-	capVal := fmt.Sprintf("%v/%v", cap.CapBytes, cap.CapPeriod.String())
-
-	// Create and md5 hex value from the parts. Note FWD and REV get different keys
-	// `fwd` TRUE if forward visa
-	// `agent` Identify agent (regardless of dock)
-	// `capID` Either the service or the group name
-	// `capVal` Expression of the value "amount for period"
-	capKey = fmt.Sprintf("%x", md5.Sum([]byte(fmt.Sprintf("%v_%v_%v_%v", fwd, clientAgentIdent, capID, capVal))))
-
-	vs.log.Debug("new data cap constraint", "capKey", capKey, "FWD", fwd, "ident", clientAgentIdent, "capID", capID)
-	rCons := vs.nodeState.ConstraintByKey(capKey)
-	if rCons == nil {
-		vs.nodeState.ProposeConstraint(&RConstraint{
-			Key:           capKey,
-			CapBytes:      cap.CapBytes,
-			PeriodSeconds: uint64(cap.CapPeriod / time.Second),
-			PeriodStarts:  uint64(time.Now().Unix()),
-		})
-		// TODO: We should wait for raft to accept the proposal.
-		remain = cap.CapBytes
-	} else {
-		// Found!
-		vs.log.Debug("data cap found", "remain", rCons.GetCapBytes())
-		pStart := time.Unix(int64(rCons.GetPeriodStarts()), 0)
-		if time.Since(pStart) > (time.Duration(rCons.PeriodSeconds) * time.Second) {
-			// period elapsed.
-			rCons.Consumed = 0
-			rCons.PeriodStarts = uint64(time.Now().Unix())
-			remain = rCons.GetCapBytes() // and so full capacity is available.
-			vs.nodeState.ProposeConstraint(rCons)
-		} else {
-			// still within a period
-			if rCons.Consumed >= rCons.CapBytes {
-				remain = 0
-			} else {
-				remain = rCons.CapBytes - rCons.Consumed
-			}
-		}
-	}
-	return
-}
-
-// ReportVisaStats is called by Docks (and forwarders?) to update visa usage details.
-func (vs *VSInst) reportVisaStats(vid uint32, capKey string, bytesUsed uint64) {
-	// Is it possible visa is removed before we get this message? Definitely.
-	vs.log.Info("visa stats", "visaID", vid, "bytesUsed", bytesUsed, "key", capKey) // TODO: add to accounting log.
-	if capKey != "" && bytesUsed > 0 {
-		if rCons := vs.nodeState.ConstraintByKey(capKey); rCons != nil {
-			// Update bytes used.
-			pStart := time.Unix(int64(rCons.GetPeriodStarts()), 0)
-			if time.Since(pStart) > (time.Duration(rCons.PeriodSeconds) * time.Second) {
-				// period elapsed.
-				rCons.Consumed = bytesUsed
-				rCons.PeriodStarts = uint64(time.Now().Unix())
-			} else {
-				// Within period
-				rCons.Consumed += bytesUsed
-			}
-			vs.nodeState.ProposeConstraint(rCons)
-		} else {
-			vs.log.Info("stats for unknown capKey, ignoring", "key", capKey)
-		}
-	}
-}
-
-// Visa will expire at the soonest of:
-//   - the agent credentials (in play) expire - (we actually just consider all creds the agent is using)
-//   - the max lifetime set in policy
-//   - the duration constraint set by the PEP in VConfig
-//   - end of datacap period, if applicable
-//
-// Note this assumes that caller has already checked the agent auth expirations.
-//
-// Also returns an explanation bitfield which indicates how the expiration was
-// computed.
-func (vs *VSInst) computeVisaExpiration(maxVisaLifetime time.Duration, durationCons, datacapPeriod time.Duration, srcAuthExp, dstAuthExp time.Time) (time.Time, libvisa.ExpFlag) {
-	var flags libvisa.ExpFlag
-	var exp time.Time
-	now := time.Now()
-	srcAuthOK := (!srcAuthExp.IsZero()) && srcAuthExp.After(now)
-	dstAuthOK := (!dstAuthExp.IsZero()) && dstAuthExp.After(now)
-	if srcAuthOK {
-		if dstAuthOK && dstAuthExp.Before(srcAuthExp) {
-			exp = dstAuthExp.Add(vs.reauthBumpTime) // give time for creds re-auth
-			flags |= libvisa.ExpFDestCreds | libvisa.ExpFBump
-		} else {
-			exp = srcAuthExp.Add(vs.reauthBumpTime) // give time for creds re-auth
-			flags |= libvisa.ExpFSrcCreds | libvisa.ExpFBump
-		}
-	} else if dstAuthOK {
-		exp = dstAuthExp.Add(vs.reauthBumpTime) // give time for creds re-auth
-		flags |= libvisa.ExpFDestCreds | libvisa.ExpFBump
-	}
-	if polExpiration := now.Add(maxVisaLifetime); exp.IsZero() || polExpiration.Before(exp) {
-		// Try using maxVisaLifetime (comes from policy global setting)
-		exp = polExpiration
-		flags |= libvisa.ExpFMaxLifetime
-	}
-	if durationCons > 0 {
-		// If there is a duration constraint on the specific policy, try that.
-		if pepEx := time.Now().Add(durationCons); exp.IsZero() || pepEx.Before(exp) {
-			exp = pepEx
-			flags |= libvisa.ExpFPolicy
-		}
-	}
-	if datacapPeriod > 0 {
-		// If there is a data cap then that will be added to the visa. The cap only applies during the period, so
-		// the visa must expire within the period.
-		if capExp := time.Now().Add(datacapPeriod); exp.IsZero() || capExp.Before(exp) {
-			exp = capExp
-			flags |= libvisa.ExpFDataCap
-		}
-	}
-	if time.Until(exp) > (35 * time.Minute) {
-		// Add some jitter so that all the visas do not bunch up
-		exp = exp.Add(time.Duration(-1*rand.Intn(30)) * time.Minute)
-		flags |= libvisa.ExpFJitter
-	}
-	return exp, flags
-}
-
-// endpointsForTraffic locate the source and destination agents by using the directory
-// to see what agent is connected at each endpoint.
-func (vs *VSInst) endpointsForTraffic(pktData *snip.Traffic) (srcAgent *agent.Agent, dstAgent *agent.Agent, err error) {
-	// Note that the visa service does not check for a route. The existence of an entry in the DirectoryService implies a route.
-	srcAgent, err = vs.AgentAtContactAddr(pktData.SrcAddr)
-	if err != nil {
-		vs.log.WithError(err).Info("visa denied: failed to resolve source ZPR address", "source", pktData.SrcAddr)
-		return nil, nil, ErrNoRouteToHost
-	}
-	dstAgent, err = vs.AgentAtContactAddr(pktData.DstAddr)
-	if err != nil {
-		vs.log.WithError(err).Info("visa denied: failed to resolve dest ZPR address", "dest", pktData.DstAddr)
-		return nil, nil, ErrNoRouteToHost
-	}
-	return
-}
-
-// inserVisaWithNewID first creates a new visa ID (based on our visa prefix, which is based on our node name),
-// then it updates the visaID field on the passed visa, and inserts the visa into our table.
-func (vs *VSInst) insertVisaWithNewID(v *vsio.Visa, isVSVisa bool, pktData *snip.Traffic) (*vtableEnt, error) {
-	vs.vtable.mtx.Lock()
-
-	// always increasing.
-	vID := vs.vtable.nextVisaID
-	if vID > maxVisaID {
-		panic(fmt.Sprintf("max visa ID reached: %d", vID)) // TODO: solve this :)
-	}
-	vs.vtable.nextVisaID = vID + 1
-	v.IssuerId = (uint32(vs.nodeNumber) << 24) | vID
-	ve := &vtableEnt{
-		v:        v,
-		isVSVisa: isVSVisa,
-		pktData:  pktData,
-	}
-	vs.vtable.table[v.IssuerId] = ve
-	sz := len(vs.vtable.table)
-	// vs.dumpVisaTableHoldingLock("insertVisa")
-	vs.vtable.mtx.Unlock()
-	vs.log.Debug("added visa", "id", vID, "isVsVisa?", isVSVisa, "netconfig", v.Configuration, "tableSize", sz)
-	return ve, nil
-}
-
-func (vs *VSInst) visaDenied(configID uint64, reason string, pktData *snip.Traffic, tetherAddr netip.Addr) {
-	if vs.vlog != nil {
-		vs.vlog.LogVisaDenied(configID, pktData, reason, tetherAddr)
-	}
-	vs.log.Info("Visa denied", "flow", pktData.Flow(), "reason", reason)
-}
-
-func (vs *VSInst) visaCreated(visa *vsio.Visa, expires time.Time, pktData *snip.Traffic, explainer string, requestor netip.Addr) {
-	if vs.vlog != nil {
-		vs.vlog.LogVisaCreated(visa, pktData, explainer, requestor)
-	}
-	vs.log.Info("visa created", "flow", pktData.Flow(), "explain", explainer, "configuration", visa.Configuration, "expires", expires)
-}
-
-// checkAndUpdateAttrs given an agent identity set and a set of auth'd attributes, first check
-// to see if any of the attributes has expired. If so run a query against the datasources
-// for updated values. The query hits our proxy/cache so may or many not actually have
-// to go all the way to the datasources.
-//
-// Note that expired claimes are removed from the returned set of attributes. Even if
-// the query to refresh them fails.
-//
-// Returns (ATTRS_CHANGED_FLAG, NEW_ATTRS, ERROR)
-//
-// If ATTRS_CHANGED_FLAG is true then the NEW_ATTRS should replace the ones on the
-// passed agent -- even if an error is indicated.
-//
-// ErrNotSupported is returned as error in case where the data source does not support
-// the Query operation.  (PROBLEM: what if there are multiple data sources?)
-func (vs *VSInst) checkAndUpdateAttrs(now time.Time, agnt *agent.Agent) (bool, map[string]*agent.ClaimV, error) {
-
-	keepAttrs := make(map[string]*agent.ClaimV)
-
-	var expired []string // we query for these
-
-	for aKey, aVx := range agnt.GetAuthedClaims() {
-		if strings.HasPrefix(aKey, "zpr.") {
-			keepAttrs[aKey] = aVx
-			continue // For now we don't know how to update zpr keys
-		}
-		if aVx.Exp.IsZero() {
-			keepAttrs[aKey] = aVx
-			continue // unset? Then it never expires.
-		}
-		if now.After(aVx.Exp) {
-			expired = append(expired, aKey)
-		} else {
-			keepAttrs[aKey] = aVx
-		}
-	}
-	if len(expired) == 0 {
-		return false, agnt.GetAuthedClaims(), nil // no changes, nothing expired.
-	}
-
-	var toks [][]byte
-	toks = append(toks, []byte(agnt.GetIdentity())) // TODO: what if there are multiple tokens on an agent?
-	qreq := &zds.QueryRequest{
-		TokenList: toks,
-		AttrKeys:  expired,
-	}
-	// Note that the keys have datasource prefixes on them.
-	// The auth service will set the prefixes on the response too
-	qresp, err := vs.attrProx.Query(now, qreq)
-	if err != nil {
-		// TODO: If datasource does not support query (eg, internal DS), should we
-		//       still remove the "expired" claims?  What is caller supposed to do?
-		//       How does caller detect when it needs to re-auth instead?
-		//
-		// FOR NOW WE ARE REMOVING EXPIRED CLAIMS !!
-		return true, keepAttrs, err
-	}
-	// Proxy may use the ttl on the response, but we do not care.
-	for _, za := range qresp.Attrs {
-		// Make sure source does not try to set any zpr keys!
-		if strings.HasPrefix(za.Key, "zpr.") {
-			vs.log.Info("invalid attempt by trusted service to set zpr key", "key", za.Key, "value", za.Val)
-			continue
-		}
-		keepAttrs[za.Key] = &agent.ClaimV{
-			V:   za.Val,
-			Exp: time.Unix(za.Exp, 0),
-		}
-	}
-	return true, keepAttrs, nil // No error, and attributes have been updated
-}
-
-// For debugging assistance only
-func (vs *VSInst) dumpVisaTableHoldingLock(reason string) {
-	vs.log.Infof("xxx starting table dump (%v) %d entries xxx", reason, len(vs.vtable.table))
-	for vid, ent := range vs.vtable.table {
-		vs.log.Info("__entry__",
-			"issuerID", vid,
-			"flow", ent.pktData.Flow(),
-			"duration", time.Until(vsio.VToTime(ent.v.Expires)).String(),
-			"successor", ent.successor)
-	}
-	vs.log.Info("xxx table dump ends xxx")
 }
